# Document/11_openapi_starter.yaml
openapi: 3.0.3
info:
  title: DoodhDirect API
  version: 1.0.0
  description: >-
    Farm-to-home dairy platform API. Application-owned business timestamps use
    India-local wall-clock values in Asia/Kolkata. Dairy business timestamps are
    accepted from clients as UTC instants and converted to India-local wall-clock
    values at the API boundary; responses serialize them without a Z suffix or
    offset. Provider/external and deferred infrastructure timestamps retain UTC
    semantics and their existing UTC-suffixed field names.
servers:
  - url: https://api.example.com/api/v1
paths:
...
  /products:
    get:
      summary: List active catalogue products
      description: Anonymous catalogue read. Results require an active product, active category, and an available active branch.
      security: []
      parameters:
        - $ref: '#/components/parameters/categoryIdQuery'
      responses:
        '200': { description: Product list envelope }
  /products/{productId}:
    get:
      summary: Get one active catalogue product
      security: []
      parameters:
        - $ref: '#/components/parameters/productId'
      responses:
        '200': { description: Product envelope }
        '404': { description: Product not found or unavailable }
...
  /admin/products:
    get:
      summary: List products for catalogue administration
      security:
        - bearerAuth: []
      responses:
        '200': { description: Product list envelope }
        '403': { description: Missing CATALOGUE.READ permission }
    post:
      summary: Create a catalogue product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertProductRequest'
      responses:
        '200': { description: Created product envelope }
        '400': { description: Validation error }
        '403': { description: Missing CATALOGUE.MANAGE permission }
        '409': { description: Duplicate SKU }
  /admin/products/{productId}:
    get:
      summary: Get one product for catalogue administration
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/productId'
      responses:
        '200': { description: Product envelope }
        '403': { description: Missing CATALOGUE.READ permission }
        '404': { description: Product not found }
    patch:
      summary: Update a catalogue product
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/productId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertProductRequest'
      responses:
        '200': { description: Updated product envelope }
        '400': { description: Validation error }
        '403': { description: Missing CATALOGUE.MANAGE permission }
        '404': { description: Product not found }
        '409': { description: Duplicate SKU }
  /admin/products/{productId}/activate:
    post:
      summary: Activate a catalogue product
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/productId'
      responses:
        '200': { description: Product envelope }
        '403': { description: Missing CATALOGUE.MANAGE permission }
        '422': { description: Active category required }
  /admin/products/{productId}/deactivate:
    post:
      summary: Deactivate a catalogue product
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/productId'
      responses:
        '200': { description: Product envelope }
        '403': { description: Missing CATALOGUE.MANAGE permission }
  /admin/products/{productId}/branches:
    put:
      summary: Set product availability for a branch
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/productId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetProductBranchAvailabilityRequest'
      responses:
        '200': { description: Updated product envelope }
        '400': { description: Validation error }
        '403': { description: Missing CATALOGUE.MANAGE permission }
        '404': { description: Product or branch not found }
  /admin/product-categories:
    get:
      summary: List all categories for catalogue administration
      security:
        - bearerAuth: []
      responses:
        '200': { description: Category list envelope }
        '403': { description: Missing CATALOGUE.READ permission }
    post:
      summary: Create a product category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertProductCategoryRequest'
      responses:
        '200': { description: Created category envelope }
        '400': { description: Validation error }
        '403': { description: Missing CATALOGUE.MANAGE permission }
        '409': { description: Duplicate category code }
  /admin/product-categories/{categoryId}:
    patch:
      summary: Update a product category
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/categoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertProductCategoryRequest'
      responses:
        '200': { description: Updated category envelope }
        '400': { description: Validation error }
        '403': { description: Missing CATALOGUE.MANAGE permission }
        '404': { description: Category not found }
  /admin/product-categories/{categoryId}/activate:
    post:
      summary: Activate a product category
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/categoryId'
      responses:
        '200': { description: Category envelope }
        '403': { description: Missing CATALOGUE.MANAGE permission }
  /admin/product-categories/{categoryId}/deactivate:
    post:
      summary: Deactivate a product category
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/categoryId'
      responses:
        '200': { description: Category envelope }
        '403': { description: Missing CATALOGUE.MANAGE permission }
...
  /admin/setup/otp-provider:
    get:
      tags: [OTP provider setup]
      summary: Read the OTP provider configuration
      description: Requires SETUP.OTP_PROVIDER.READ (OWNER and SYSTEM_ADMIN only). The MSG91 AuthKey is write-only and is never returned.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OTP provider configuration envelope
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OtpProviderConfigurationResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    put:
      tags: [OTP provider setup]
      summary: Update the OTP provider configuration
      description: Requires SETUP.OTP_PROVIDER.MANAGE (OWNER and SYSTEM_ADMIN only). A non-empty AuthKey is encrypted at rest with IDataProtector before persistence and never returned by reads or audits.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateOtpProviderConfigurationRequest' }
      responses:
        '200':
          description: Updated OTP provider configuration envelope
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OtpProviderConfigurationResult' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/setup/otp-provider/test:
    post:
      tags: [OTP provider setup]
      summary: Test the OTP provider configuration
      description: Requires SETUP.OTP_PROVIDER.MANAGE (OWNER and SYSTEM_ADMIN only). Sends a provider probe and reports success or a safe diagnostic message. Requires the provider to be fully configured and enabled.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OTP provider test result envelope
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OtpProviderTestResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/BusinessRuleError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    categoryIdQuery:
      name: categoryId
      in: query
      required: false
      schema: { type: string, format: uuid }
    productId:
      name: productId
      in: path
      required: true
      schema: { type: string, format: uuid }
    categoryId:
      name: categoryId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Authentication is required or the access token is invalid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: The authenticated user lacks the required permission
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    BusinessRuleError:
      description: A subscription, entitlement, cutoff, availability, or payment business rule failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        field: { type: string, nullable: true }
        message: { type: string }
    ErrorResponse:
      type: object
      required: [success, errors]
      properties:
        success: { type: boolean, enum: [false] }
        data: { nullable: true }
        message: { type: string, nullable: true }
        errors:
          type: array
          items: { $ref: '#/components/schemas/ApiError' }
...
    OtpProviderConfigurationResult:
      type: object
      required: [provider, enabled, configured, status]
      properties:
        provider: { type: string, description: OTP provider code (MSG91). }
        enabled: { type: boolean }
        widgetId: { type: string, nullable: true }
        environment: { type: string, nullable: true, enum: [Test, Production], description: MSG91 environment. }
        configured: { type: boolean, description: True when both a widget id and an auth key are present. }
        status: { type: string, description: Human-readable configuration status. }
    UpdateOtpProviderConfigurationRequest:
      type: object
      properties:
        enabled: { type: boolean, nullable: true }
        widgetId: { type: string, nullable: true }
        authKey: { type: string, nullable: true, format: password, description: MSG91 auth key. Write-only; encrypted at rest with IDataProtector and never returned. An empty value is rejected. }
        environment: { type: string, nullable: true, enum: [Test, Production] }
    OtpProviderTestResult:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean }
        message: { type: string, description: Safe diagnostic message; never contains credentials. }
...
    UpsertProductRequest:
      type: object
      required: [sku, name, categoryId, unitOfMeasure, price]
      properties:
        sku: { type: string, maxLength: 80 }
        name: { type: string, maxLength: 160 }
        description: { type: string, nullable: true, maxLength: 1000 }
        categoryId: { type: string, format: uuid }
        unitOfMeasure: { type: string, enum: [litre, kilogram, gram, piece] }
        price: { type: number, format: double, exclusiveMinimum: 0, multipleOf: 0.01 }
    UpsertProductCategoryRequest:
      type: object
      required: [code, name]
      properties:
        code: { type: string, maxLength: 40 }
        name: { type: string, maxLength: 120 }
        description: { type: string, nullable: true, maxLength: 500 }
    SetProductBranchAvailabilityRequest:
      type: object
      required: [branchId, isAvailable]
      properties:
        branchId: { type: string, format: uuid }
        isAvailable: { type: boolean }
        maxDailyQuantity: { type: number, format: double, nullable: true, exclusiveMinimum: 0, multipleOf: 0.001 }
...